# Switch from os.walk to scandir.walk to improve performance