@ppolewicz 2019-08-02
The test above has shown an improvement by a factor of 66.949 on 20k files, but I later ran the same test on 200k files to isolate it from startup time.
Result using version before this PR:
```bash
real	75m3,426s
user	0m5,790s
sys	0m13,213s
```
and after PR:
```bash
real	0m2,630s
user	0m2,083s
sys	0m0,123s
```
so the speedup is not ~67 times but ~1712 times, when considering flat directories.